% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2bibtexR.R
\name{pmid2bibtexR}
\alias{pmid2bibtexR}
\title{pmid2bibtexR}
\usage{
pmid2bibtexR(pmid = c(20364295, 21146710), n_author = "all")
}
\arguments{
\item{pmid}{list of pubmed PMID}

\item{n_author}{Number of Authors to see in BibTeX format. Default to "all"}
}
\description{
This function allows the user to get BibTeX format from PMID using Entrez from biopython.
}
\examples{
pmid <- c(21146710, 20364295)
bib <- pmid2bibtexR(pmid, n_author = "all")
}
