% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid2bibtexR.R
\name{pmid2bibtexR}
\alias{pmid2bibtexR}
\title{pmid2bibtexR}
\usage{
pmid2bibtexR(
  pmid,
  myemail = "user@example.com",
  myapikey = NULL,
  n_author = "all"
)
}
\arguments{
\item{pmid}{list of pubmed PMID}

\item{myemail}{Bio.Entrez needs e-mail address of user.}

\item{myapikey}{Pubmed allows only 3 requests per second without API Key. Default to None. 10 requests per second with API Key. See the details: https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/}

\item{n_author}{Number of Authors to see in BibTeX format. Default to "all"}
}
\description{
This function allows the user to get BibTeX format from PMID using Entrez from biopython.
}
\examples{
pmid <- c(33538338, 32682809)
pmid2bibtexR(pmid, myemail = "user@example.com", n_author = "all")
}
